backend/pyproject.toml
[project]
name = "llm-tic-tac-toe"
version = "0.1.0"
description = "LLMs play tic-tac-toe"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
